Refactor perfmon counter enums into oshi-common
Move performance counter enum definitions from oshi-core driver classes
into oshi.driver.common.windows.perfmon (oshi-common) so they are
available to both JNA and FFM implementations.

- Extract PdhCounterProperty and PdhCounterWildcardProperty as
  standalone interfaces in oshi-common
- Move instance filter constants (TOTAL_INSTANCE, etc.) and GPU
  constants into PerfmonConstants
- Create enum-only classes in oshi-common: ProcessInformation,
  ProcessorInformation, MemoryInformation, PagingFile, PhysicalDisk,
  SystemInformation, ThreadInformation, GpuInformation
- Rename oshi-core driver classes to *JNA, retaining only the static
  query methods
- Remove nested interfaces from PerfCounterQuery and
  PerfCounterWildcardQuery
- Update all callers to import enums from oshi-common
