+++ oshi-common/src/main/java/oshi/driver/common/windows/perfmon/GpuInformation.java
@@ -2,25 +2,15 @@
  * Copyright 2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.driver.windows.perfmon;
-
-import java.util.List;
-import java.util.Map;
+package oshi.driver.common.windows.perfmon;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.util.platform.windows.PerfCounterWildcardQuery;
-import oshi.util.platform.windows.PerfCounterWildcardQuery.PdhCounterWildcardProperty;
-import oshi.util.tuples.Pair;
 
 /**
- * Utility to query GPU Engine and GPU Adapter Memory performance counters. Available on Windows 10 version 1709 and
- * later.
+ * GPU performance counter enums. Available on Windows 10 version 1709 and later.
  */
 @ThreadSafe
 public final class GpuInformation {
-
-    static final String GPU_ENGINE = "GPU Engine";
-    static final String GPU_ADAPTER_MEMORY = "GPU Adapter Memory";
 
     /**
      * GPU Engine running time counter properties. Instance names have the form:
@@ -66,24 +56,5 @@ public final class GpuInformation {
     }
 
     private GpuInformation() {
-    }
-
-    /**
-     * Queries GPU Engine running time counters for all instances.
-     *
-     * @return pair of instance name list and counter value map
-     */
-    public static Pair<List<String>, Map<GpuEngineProperty, List<Long>>> queryGpuEngineCounters() {
-        return PerfCounterWildcardQuery.queryInstancesAndValuesFromPDH(GpuEngineProperty.class, GPU_ENGINE);
-    }
-
-    /**
-     * Queries GPU Adapter Memory counters for all instances.
-     *
-     * @return pair of instance name list and counter value map
-     */
-    public static Pair<List<String>, Map<GpuAdapterMemoryProperty, List<Long>>> queryGpuAdapterMemoryCounters() {
-        return PerfCounterWildcardQuery.queryInstancesAndValuesFromPDH(GpuAdapterMemoryProperty.class,
-                GPU_ADAPTER_MEMORY);
     }
 }
