+++ oshi-common/src/main/java/oshi/driver/common/windows/perfmon/MemoryInformation.java
@@ -2,20 +2,12 @@
  * Copyright 2020-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.driver.windows.perfmon;
-
-import static oshi.driver.common.windows.perfmon.PerfmonConstants.MEMORY;
-import static oshi.driver.common.windows.perfmon.PerfmonConstants.WIN32_PERF_RAW_DATA_PERF_OS_MEMORY;
-
-import java.util.Collections;
-import java.util.Map;
+package oshi.driver.common.windows.perfmon;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.util.platform.windows.PerfCounterQuery;
-import oshi.util.platform.windows.PerfCounterQuery.PdhCounterProperty;
 
 /**
- * Utility to query Memory performance counter
+ * Memory performance counter enums
  */
 @ThreadSafe
 public final class MemoryInformation {
@@ -47,17 +39,5 @@ public final class MemoryInformation {
     }
 
     private MemoryInformation() {
-    }
-
-    /**
-     * Returns page swap counters
-     *
-     * @return Page swap counters for memory.
-     */
-    public static Map<PageSwapProperty, Long> queryPageSwaps() {
-        if (PerfmonDisabled.PERF_OS_DISABLED) {
-            return Collections.emptyMap();
-        }
-        return PerfCounterQuery.queryValues(PageSwapProperty.class, MEMORY, WIN32_PERF_RAW_DATA_PERF_OS_MEMORY);
     }
 }
