+++ oshi-common/src/main/java/oshi/driver/common/windows/perfmon/ThreadInformation.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2020-2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+package oshi.driver.common.windows.perfmon;
+
+import static oshi.driver.common.windows.perfmon.PerfmonConstants.NOT_TOTAL_INSTANCES;
+
+import oshi.annotation.concurrent.ThreadSafe;
+
+/**
+ * Thread performance counter enums
+ */
+@ThreadSafe
+public final class ThreadInformation {
+
+    /**
+     * Thread performance counters
+     */
+    public enum ThreadPerformanceProperty implements PdhCounterWildcardProperty {
+        // First element defines WMI instance name field and PDH instance filter
+        NAME(NOT_TOTAL_INSTANCES),
+        // Remaining elements define counters
+        PERCENTUSERTIME("% User Time"), //
+        PERCENTPRIVILEGEDTIME("% Privileged Time"), //
+        ELAPSEDTIME("Elapsed Time"), //
+        PRIORITYCURRENT("Priority Current"), //
+        STARTADDRESS("Start Address"), //
+        THREADSTATE("Thread State"), //
+        THREADWAITREASON("Thread Wait Reason"), // 5 is SUSPENDED
+        IDPROCESS("ID Process"), //
+        IDTHREAD("ID Thread"), //
+        CONTEXTSWITCHESPERSEC("Context Switches/sec");
+
+        private final String counter;
+
+        ThreadPerformanceProperty(String counter) {
+            this.counter = counter;
+        }
+
+        @Override
+        public String getCounter() {
+            return counter;
+        }
+    }
+
+    private ThreadInformation() {
+    }
+}
