+++ oshi-core/src/main/java/oshi/driver/windows/perfmon/PagingFileJNA.java
@@ -11,41 +11,16 @@ import java.util.Collections;
 import java.util.Map;
 
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.driver.common.windows.perfmon.PagingFile.PagingPercentProperty;
 import oshi.util.platform.windows.PerfCounterQuery;
-import oshi.util.platform.windows.PerfCounterQuery.PdhCounterProperty;
 
 /**
  * Utility to query Paging File performance counter
  */
 @ThreadSafe
-public final class PagingFile {
+public final class PagingFileJNA {
 
-    /**
-     * For swap file usage
-     */
-    public enum PagingPercentProperty implements PdhCounterProperty {
-        PERCENTUSAGE(PerfCounterQuery.TOTAL_INSTANCE, "% Usage");
-
-        private final String instance;
-        private final String counter;
-
-        PagingPercentProperty(String instance, String counter) {
-            this.instance = instance;
-            this.counter = counter;
-        }
-
-        @Override
-        public String getInstance() {
-            return instance;
-        }
-
-        @Override
-        public String getCounter() {
-            return counter;
-        }
-    }
-
-    private PagingFile() {
+    private PagingFileJNA() {
     }
 
     /**
