+++ oshi-core/src/main/java/oshi/driver/windows/perfmon/ThreadInformationJNA.java
@@ -13,48 +13,17 @@ import java.util.Locale;
 import java.util.Map;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.util.platform.windows.PerfCounterQuery;
+import oshi.driver.common.windows.perfmon.ThreadInformation.ThreadPerformanceProperty;
 import oshi.util.platform.windows.PerfCounterWildcardQuery;
-import oshi.util.platform.windows.PerfCounterWildcardQuery.PdhCounterWildcardProperty;
 import oshi.util.tuples.Pair;
 
 /**
  * Utility to query Thread Information performance counter
  */
 @ThreadSafe
-public final class ThreadInformation {
+public final class ThreadInformationJNA {
 
-    /**
-     * Thread performance counters
-     */
-    public enum ThreadPerformanceProperty implements PdhCounterWildcardProperty {
-        // First element defines WMI instance name field and PDH instance filter
-        NAME(PerfCounterQuery.NOT_TOTAL_INSTANCES),
-        // Remaining elements define counters
-        PERCENTUSERTIME("% User Time"), //
-        PERCENTPRIVILEGEDTIME("% Privileged Time"), //
-        ELAPSEDTIME("Elapsed Time"), //
-        PRIORITYCURRENT("Priority Current"), //
-        STARTADDRESS("Start Address"), //
-        THREADSTATE("Thread State"), //
-        THREADWAITREASON("Thread Wait Reason"), // 5 is SUSPENDED
-        IDPROCESS("ID Process"), //
-        IDTHREAD("ID Thread"), //
-        CONTEXTSWITCHESPERSEC("Context Switches/sec");
-
-        private final String counter;
-
-        ThreadPerformanceProperty(String counter) {
-            this.counter = counter;
-        }
-
-        @Override
-        public String getCounter() {
-            return counter;
-        }
-    }
-
-    private ThreadInformation() {
+    private ThreadInformationJNA() {
     }
 
     /**
