+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsCentralProcessor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2025 The OSHI Project Contributors
+ * Copyright 2016-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 package oshi.hardware.platform.windows;
@@ -20,24 +20,24 @@ import org.slf4j.LoggerFactory;
 
 import com.sun.jna.Native;
 import com.sun.jna.platform.win32.Advapi32Util;
+import com.sun.jna.platform.win32.COM.WbemcliUtil.WmiResult;
 import com.sun.jna.platform.win32.PowrProf.POWER_INFORMATION_LEVEL;
 import com.sun.jna.platform.win32.VersionHelpers;
 import com.sun.jna.platform.win32.Win32Exception;
 import com.sun.jna.platform.win32.WinBase;
 import com.sun.jna.platform.win32.WinReg;
-import com.sun.jna.platform.win32.COM.WbemcliUtil.WmiResult;
 
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.driver.common.windows.perfmon.ProcessorInformation.InterruptsProperty;
+import oshi.driver.common.windows.perfmon.ProcessorInformation.ProcessorFrequencyProperty;
+import oshi.driver.common.windows.perfmon.ProcessorInformation.ProcessorTickCountProperty;
+import oshi.driver.common.windows.perfmon.ProcessorInformation.ProcessorUtilityTickCountProperty;
+import oshi.driver.common.windows.perfmon.ProcessorInformation.SystemTickCountProperty;
+import oshi.driver.common.windows.perfmon.SystemInformation.ContextSwitchProperty;
 import oshi.driver.windows.LogicalProcessorInformation;
 import oshi.driver.windows.perfmon.LoadAverage;
-import oshi.driver.windows.perfmon.ProcessorInformation;
-import oshi.driver.windows.perfmon.ProcessorInformation.InterruptsProperty;
-import oshi.driver.windows.perfmon.ProcessorInformation.ProcessorFrequencyProperty;
-import oshi.driver.windows.perfmon.ProcessorInformation.ProcessorTickCountProperty;
-import oshi.driver.windows.perfmon.ProcessorInformation.ProcessorUtilityTickCountProperty;
-import oshi.driver.windows.perfmon.ProcessorInformation.SystemTickCountProperty;
-import oshi.driver.windows.perfmon.SystemInformation;
-import oshi.driver.windows.perfmon.SystemInformation.ContextSwitchProperty;
+import oshi.driver.windows.perfmon.ProcessorInformationJNA;
+import oshi.driver.windows.perfmon.SystemInformationJNA;
 import oshi.driver.windows.wmi.Win32Processor;
 import oshi.driver.windows.wmi.Win32Processor.ProcessorIdProperty;
 import oshi.hardware.common.AbstractCentralProcessor;
@@ -219,7 +219,7 @@ final class WindowsCentralProcessor extends AbstractCentralProcessor {
             // IRQ and ticks:
             // Percent time raw value is cumulative 100NS-ticks
             // Divide by 10_000 to get milliseconds
-            Map<SystemTickCountProperty, Long> valueMap = ProcessorInformation.querySystemCounters();
+            Map<SystemTickCountProperty, Long> valueMap = ProcessorInformationJNA.querySystemCounters();
             ticks[TickType.IRQ.getIndex()] = valueMap.getOrDefault(SystemTickCountProperty.PERCENTINTERRUPTTIME, 0L)
                     / 10_000L;
             ticks[TickType.SOFTIRQ.getIndex()] = valueMap.getOrDefault(SystemTickCountProperty.PERCENTDPCTIME, 0L)
@@ -252,7 +252,7 @@ final class WindowsCentralProcessor extends AbstractCentralProcessor {
     @Override
     public long[] queryCurrentFreq() {
         if (VersionHelpers.IsWindows7OrGreater()) {
-            Pair<List<String>, Map<ProcessorFrequencyProperty, List<Long>>> instanceValuePair = ProcessorInformation
+            Pair<List<String>, Map<ProcessorFrequencyProperty, List<Long>>> instanceValuePair = ProcessorInformationJNA
                     .queryFrequencyCounters();
             List<String> instances = instanceValuePair.getA();
             Map<ProcessorFrequencyProperty, List<Long>> valueMap = instanceValuePair.getB();
@@ -368,7 +368,7 @@ final class WindowsCentralProcessor extends AbstractCentralProcessor {
             initProcessorUtilityBase = initialUtilityCounters
                     .get(ProcessorUtilityTickCountProperty.PERCENTPROCESSORUTILITY_BASE);
         } else {
-            Pair<List<String>, Map<ProcessorTickCountProperty, List<Long>>> instanceValuePair = ProcessorInformation
+            Pair<List<String>, Map<ProcessorTickCountProperty, List<Long>>> instanceValuePair = ProcessorInformationJNA
                     .queryProcessorCounters();
             instances = instanceValuePair.getA();
             Map<ProcessorTickCountProperty, List<Long>> valueMap = instanceValuePair.getB();
@@ -496,17 +496,17 @@ final class WindowsCentralProcessor extends AbstractCentralProcessor {
     }
 
     private static Pair<List<String>, Map<ProcessorUtilityTickCountProperty, List<Long>>> queryProcessorUtilityCounters() {
-        return ProcessorInformation.queryProcessorCapacityCounters();
+        return ProcessorInformationJNA.queryProcessorCapacityCounters();
     }
 
     @Override
     public long queryContextSwitches() {
-        return SystemInformation.queryContextSwitchCounters().getOrDefault(ContextSwitchProperty.CONTEXTSWITCHESPERSEC,
-                0L);
+        return SystemInformationJNA.queryContextSwitchCounters()
+                .getOrDefault(ContextSwitchProperty.CONTEXTSWITCHESPERSEC, 0L);
     }
 
     @Override
     public long queryInterrupts() {
-        return ProcessorInformation.queryInterruptCounters().getOrDefault(InterruptsProperty.INTERRUPTSPERSEC, 0L);
+        return ProcessorInformationJNA.queryInterruptCounters().getOrDefault(InterruptsProperty.INTERRUPTSPERSEC, 0L);
     }
 }
