+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsGpuStats.java
@@ -15,9 +15,9 @@ import org.slf4j.LoggerFactory;
 import com.sun.jna.platform.win32.COM.WbemcliUtil.WmiResult;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.driver.windows.perfmon.GpuInformation;
-import oshi.driver.windows.perfmon.GpuInformation.GpuAdapterMemoryProperty;
-import oshi.driver.windows.perfmon.GpuInformation.GpuEngineProperty;
+import oshi.driver.common.windows.perfmon.GpuInformation.GpuAdapterMemoryProperty;
+import oshi.driver.common.windows.perfmon.GpuInformation.GpuEngineProperty;
+import oshi.driver.windows.perfmon.GpuInformationJNA;
 import oshi.driver.windows.wmi.LhmSensor;
 import oshi.driver.windows.wmi.LhmSensor.LhmSensorProperty;
 import oshi.hardware.GpuStats;
@@ -95,7 +95,7 @@ final class WindowsGpuStats implements GpuStats {
         if (luidPrefix.isEmpty()) {
             return new GpuTicks(0L, 0L);
         }
-        Pair<List<String>, Map<GpuEngineProperty, List<Long>>> engineData = GpuInformation.queryGpuEngineCounters();
+        Pair<List<String>, Map<GpuEngineProperty, List<Long>>> engineData = GpuInformationJNA.queryGpuEngineCounters();
         List<String> instances = engineData.getA();
         Map<GpuEngineProperty, List<Long>> values = engineData.getB();
         List<Long> runningTimes = values.get(GpuEngineProperty.RUNNING_TIME);
@@ -311,7 +311,7 @@ final class WindowsGpuStats implements GpuStats {
         if (luidPrefix.isEmpty()) {
             return -1L;
         }
-        Pair<List<String>, Map<GpuAdapterMemoryProperty, List<Long>>> adapterData = GpuInformation
+        Pair<List<String>, Map<GpuAdapterMemoryProperty, List<Long>>> adapterData = GpuInformationJNA
                 .queryGpuAdapterMemoryCounters();
         List<String> instances = adapterData.getA();
         List<Long> values = adapterData.getB().get(property);
